requirements.txt
# Core dependencies
rank-bm25>=0.2.2
python-frontmatter>=1.1.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0

# CLI framework
click>=8.1.7

# LLM integration
ollama>=0.1.0

# Voice processing (Whisper)
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
scipy>=1.10.0
numpy>=1.24.0
torch>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
